requirements.txt
PyGithub==1.59.1
python-dotenv>=1.0.0   # If using .env files to store the token
requests>=2.31.0       # If using raw GitHub API calls
openai